# stats/target/classes/openapi.yaml
openapi: 3.1.0
info:
  title: MetaMapa - Servicio de Estadísticas
  version: 1.0.0
  description: >
    Microservicio de solo lectura que calcula y expone estadísticas agregadas
    (provincias, categorías, distribución horaria y spam) y permite exportarlas en CSV.
    Los datos se recalculan periódicamente mediante un scheduler y se persisten en el
    repositorio local del servicio.
servers:
  - url: https://api.tu-dominio.org/api
    description: Producción
  - url: https://staging.tu-dominio.org/api
    description: Staging
...
security:
  - ApiKeyAuth: []
  # - BearerAuth: []  # Si preferís JWT, habilitá esta línea y deshabilitá ApiKeyAuth arriba.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponseSimple:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        details:
          type: object
          additionalProperties: true

    # ---------- DTOs ----------
...
    CSVExportDTO:
      type: object
      properties:
        tipoExportacion: { type: string, example: "provincias" }
        nombreArchivo: { type: string, example: "provincias_123.csv" }
        rutaArchivo: { type: string, example: "/exports/2025-10-30/provincias_123.csv" }
        fechaExportacion: { type: string, format: date-time, example: "2025-10-30T11:05:00Z" }
        totalRegistros: { type: integer, format: int64, example: 24 }
        columnas:
          type: array
          items: { type: string }
          example: [ "coleccionId", "provincia", "cantidadHechos", "porcentaje", "fechaCalculo" ]
        estado:
          type: string
          enum: [ COMPLETADO, EN_PROCESO, ERROR ]
          example: COMPLETADO
        mensaje: { type: string, example: "Exportación generada correctamente" }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string, example: "Bad Request" }
        message: { type: string, example: "Parámetro 'categoria' es requerido" }
        timestamp: { type: string, format: date-time }
        path: { type: string, example: "/api/stats/hora/categoria/Incendio%20Forestal" }
  parameters:
    ColeccionIdPath:
      name: coleccionId
      in: path
      required: true
      description: Identificador numérico de la colección
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Parámetros inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
paths:
...
  /stats/export/provincias/{coleccionId}:
    get:
      tags: [Export]
      summary: Exporta CSV de provincias para una colección
      parameters:
        - $ref: '#/components/parameters/ColeccionIdPath'
      responses:
        '200':
          description: CSV generado
          headers:
            Content-Disposition:
              description: Nombre sugerido de archivo
              schema: { type: string, example: 'attachment; filename="provincias_123.csv"' }
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/export/categorias:
    get:
      tags: [Export]
      summary: Exporta CSV de categorías
      responses:
        '200':
          description: CSV generado
          headers:
            Content-Disposition:
              description: Nombre sugerido de archivo
              schema: { type: string, example: 'attachment; filename="categorias.csv"' }
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/export/completo:
    get:
      tags: [Export]
      summary: Exporta CSV consolidado con todas las métricas
      responses:
        '200':
          description: CSV generado
          headers:
            Content-Disposition:
              description: Nombre sugerido de archivo
              schema: { type: string, example: 'attachment; filename="estadisticas_completo.csv"' }
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/export/limpiar:
    post:
      tags: [Export]
      summary: Limpia archivos CSV antiguos (housekeeping)
      responses:
        '200':
          description: Operación ejecutada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSimple' }
        '500': { $ref: '#/components/responses/InternalError' }
  /stats/export/archivos:
    get:
      tags: [Export]
      summary: Lista exportaciones CSV disponibles
      responses:
        '200':
          description: Listado de archivos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CSVExportDTO' }
                  message: { type: string }
        '500': { $ref: '#/components/responses/InternalError' }

  # ====================== ADMIN ======================
  /admin/stats/recalcular:
    post:
      tags: [Admin]
      summary: Fuerza recálculo de estadísticas
      responses:
        '200':
          description: Recalculo iniciado/completado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSimple' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/stats/cache/limpiar:
    post:
      tags: [Admin]
      summary: Limpia caché interno de estadísticas (si aplica)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSimple' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/stats/estado:
    get:
      tags: [Admin]
      summary: Estado del servicio (última corrida, versión, etc.)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    additionalProperties: true
                    example:
                      ultimaEjecucion: "2025-10-30T11:00:00Z"
                      version: "1.0.0"
                      scheduler: "0 0 * * *"
                  message: { type: string }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/stats/csv/limpiar:
    post:
      tags: [Admin]
      summary: Limpia exportaciones CSV (equivalente operativo)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseSimple' }
        '500': { $ref: '#/components/responses/InternalError' }